tidy up configs
