+++ models/tt_transformers/tt/common.py
@@ -4,9 +4,12 @@
 
 import math
 import re
+from enum import Enum
+from typing import Optional
 
 import torch
 from loguru import logger
+from pydantic import BaseModel, Field
 
 import ttnn
 
@@ -25,6 +28,50 @@ class PagedAttentionConfig:
     def __init__(self, block_size=32, max_num_blocks=1024):
         self.block_size = block_size
         self.max_num_blocks = max_num_blocks
+
+
+class RopeScalingType(str, Enum):
+    """Types of RoPE scaling."""
+
+    # LINEAR = "linear"
+    # DYNAMIC = "dynamic"
+    YARN = "yarn"
+    LLAMA3 = "llama3"
+
+
+class RopeScaling(BaseModel):
+    """RoPE scaling configuration."""
+
+    rope_type: RopeScalingType = Field(exclude=True, description="RoPE scaling type")
+    factor: float
+    original_max_position_embeddings: int
+
+
+class RopeScalingLlama3(RopeScaling):
+    """RoPE scaling configuration for Llama-3.x."""
+
+    # Llama-3.x specific parameters
+    low_freq_factor: Optional[float] = 1.0
+    high_freq_factor: Optional[float] = 4.0
+
+
+class RopeScalingYarn(RopeScaling):
+    """RoPE scaling configuration for Yarn."""
+
+    # Yarn-specific parameters - we could have a separate class for each type
+    beta_fast: Optional[int] = 32
+    beta_slow: Optional[int] = 1
+    mscale: Optional[float] = 1.0
+    mscale_all_dim: Optional[float] = 0.0
+
+
+def rope_scaling_model_factory(rope_scaling_params: dict) -> RopeScaling:
+    if rope_scaling_params.get("rope_type") == RopeScalingType.LLAMA3:
+        return RopeScalingLlama3(**rope_scaling_params)
+    elif rope_scaling_params.get("rope_type") == RopeScalingType.YARN or rope_scaling_params.get("rope_type") == "yarn":
+        return RopeScalingYarn(**rope_scaling_params)
+    else:
+        raise ValueError(f"Invalid RoPE scaling type: {rope_scaling_params.get('rope_type')}")
 
 
 def encode_prompt_instruct(tokenizer, prompt_text, system_prompt_text=None):
